added some improvements for validation part in signup form
